requirements.txt
fastapi==0.111.1
uvicorn==0.25.2
pydantic==2.11.2
numpy==1.26.2
sounddevice==0.4.10
twilio==8.15.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.8
openai==0.27.10
litellm==0.3.2
openai-agents
openai-agents[voice]